Cleanup of optimizer
-merge utils/lr_policy.py into optimizer
-construct_optimizer() refactored
