+++ pycls/core/optimizer.py
@@ -7,7 +7,7 @@
 
 """Optimizer."""
 
-import pycls.utils.lr_policy as lr_policy
+import numpy as np
 import torch
 from pycls.core.config import cfg
 
@@ -33,31 +33,16 @@ def construct_optimizer(model):
     when the learning rate is changed there is no need to perform the
     momentum correction by scaling V (unlike in the Caffe2 case).
     """
-    # Batchnorm parameters.
-    bn_params = []
-    # Non-batchnorm parameters.
-    non_bn_parameters = []
-    for name, p in model.named_parameters():
-        if "bn" in name:
-            bn_params.append(p)
-        else:
-            non_bn_parameters.append(p)
-    # Apply different weight decay to Batchnorm and non-batchnorm parameters.
-    bn_weight_decay = (
-        cfg.BN.CUSTOM_WEIGHT_DECAY
-        if cfg.BN.USE_CUSTOM_WEIGHT_DECAY
-        else cfg.OPTIM.WEIGHT_DECAY
-    )
-    optim_params = [
-        {"params": bn_params, "weight_decay": bn_weight_decay},
-        {"params": non_bn_parameters, "weight_decay": cfg.OPTIM.WEIGHT_DECAY},
-    ]
-    # Check all parameters will be passed into optimizer.
-    assert len(list(model.parameters())) == len(non_bn_parameters) + len(
-        bn_params
-    ), "parameter size does not match: {} + {} != {}".format(
-        len(non_bn_parameters), len(bn_params), len(list(model.parameters()))
-    )
+    if cfg.BN.USE_CUSTOM_WEIGHT_DECAY:
+        # Apply different weight decay to Batchnorm and non-batchnorm parameters.
+        p_bn = [p for n, p in model.named_parameters() if "bn" in n]
+        p_non_bn = [p for n, p in model.named_parameters() if "bn" not in n]
+        optim_params = [
+            {"params": p_bn, "weight_decay": cfg.BN.CUSTOM_WEIGHT_DECAY},
+            {"params": p_non_bn, "weight_decay": cfg.OPTIM.WEIGHT_DECAY},
+        ]
+    else:
+        optim_params = model.parameters()
     return torch.optim.SGD(
         optim_params,
         lr=cfg.OPTIM.BASE_LR,
@@ -68,9 +53,40 @@ def construct_optimizer(model):
     )
 
 
+def lr_fun_steps(cur_epoch):
+    """Steps schedule (cfg.OPTIM.LR_POLICY = 'steps')."""
+    ind = [i for i, s in enumerate(cfg.OPTIM.STEPS) if cur_epoch >= s][-1]
+    return cfg.OPTIM.BASE_LR * (cfg.OPTIM.LR_MULT ** ind)
+
+
+def lr_fun_exp(cur_epoch):
+    """Exponential schedule (cfg.OPTIM.LR_POLICY = 'exp')."""
+    return cfg.OPTIM.BASE_LR * (cfg.OPTIM.GAMMA ** cur_epoch)
+
+
+def lr_fun_cos(cur_epoch):
+    """Cosine schedule (cfg.OPTIM.LR_POLICY = 'cos')."""
+    base_lr, max_epoch = cfg.OPTIM.BASE_LR, cfg.OPTIM.MAX_EPOCH
+    return 0.5 * base_lr * (1.0 + np.cos(np.pi * cur_epoch / max_epoch))
+
+
+def get_lr_fun():
+    """Retrieves the specified lr policy function"""
+    lr_fun = "lr_fun_" + cfg.OPTIM.LR_POLICY
+    if lr_fun not in globals():
+        raise NotImplementedError("Unknown LR policy:" + cfg.OPTIM.LR_POLICY)
+    return globals()[lr_fun]
+
+
 def get_epoch_lr(cur_epoch):
-    """Retrieves the lr for the given epoch (as specified by the lr policy)."""
-    return lr_policy.get_epoch_lr(cur_epoch)
+    """Retrieves the lr for the given epoch according to the policy."""
+    lr = get_lr_fun()(cur_epoch)
+    # Linear warmup
+    if cur_epoch < cfg.OPTIM.WARMUP_EPOCHS:
+        alpha = cur_epoch / cfg.OPTIM.WARMUP_EPOCHS
+        warmup_factor = cfg.OPTIM.WARMUP_FACTOR * (1.0 - alpha) + alpha
+        lr *= warmup_factor
+    return lr
 
 
 def set_lr(optimizer, new_lr):
